Fix maav-camera CameraDriverHelper creation to not create disabled camera interfaces. Fix librealsense install script to correctly patch kernel modules. Make t265 work in camera driver despite major bug in realsense and document the existence of the bug for when intel fixes it.

+++ scripts/tools/generate_zcm_messages.sh
@@ -10,6 +10,8 @@ ZCM_TYPES_DIR="${SOFTWARE_DIR}/msgtypes/zcm"
 echo ${SOFTWARE_DIR}
 
 # Delete Generated messages
-rm -r ${SOFTWARE_DIR}/generated/common
+if [ -d "${SOFTWARE_DIR}/generated/common" ]; then
+    rm -r ${SOFTWARE_DIR}/generated/common
+fi
 
 zcm-gen --cpp --cpp-hpath ${GENERATED_DIR} ${ZCM_TYPES_DIR}/*.zcm
